refactor: update sdk providers to hooks
