+++ src/containers/Tables/BidTable/index.tsx
@@ -1,7 +1,7 @@
 import { useWeb3React } from '@web3-react/core';
 import { ethers } from 'ethers';
 import { useBidProvider } from 'lib/hooks/useBidProvider';
-import { useZnsSdk } from 'lib/providers/ZnsSdkProvider';
+import { useZnsSdk } from 'lib/hooks/sdk';
 import { useEffect, useRef, useState } from 'react';
 import { BigNumber } from 'ethers';
