+++ src/containers/Tables/OwnedDomainsTable/hooks/useOwnedDomains.ts
@@ -1,7 +1,7 @@
 import { Domain } from '@zero-tech/zns-sdk';
 import { useDidMount } from 'lib/hooks/useDidMount';
 import { useUpdateEffect } from 'lib/hooks/useUpdateEffect';
-import { useZnsSdk } from 'lib/providers/ZnsSdkProvider';
+import { useZnsSdk } from 'lib/hooks/sdk';
 import { useRef, useState } from 'react';
 
 type UseOwnedDomainsReturn = {
