+++ src/containers/Tables/SubdomainTable/hooks/useSubdomainTableData.ts
@@ -1,6 +1,6 @@
 import { DomainMetricsCollection } from '@zero-tech/zns-sdk';
 import { useUpdateEffect } from 'lib/hooks/useUpdateEffect';
-import { useZnsSdk } from 'lib/providers/ZnsSdkProvider';
+import { useZnsSdk } from 'lib/hooks/sdk';
 import { DisplayDomain } from 'lib/types';
 import { useState } from 'react';
 import useAsyncEffect from 'use-async-effect';
