+++ src/pages/ZNS/ZNS.tsx
@@ -17,7 +17,6 @@ import { SubdomainTable, CurrentDomainPreview, Raffle } from 'containers';
 //- Library Imports
 import { NFTView, TransferOwnership } from 'containers';
 import { useCurrentDomain } from 'lib/providers/CurrentDomainProvider';
-import { useZnsSdk } from 'lib/providers/ZnsSdkProvider';
 import { DomainMetrics } from '@zero-tech/zns-sdk/lib/types';
 import { ethers } from 'ethers';
 import useCurrency from 'lib/hooks/useCurrency';
@@ -26,6 +25,7 @@ import useScrollDetection from 'lib/hooks/useScrollDetection';
 import { useDidMount } from 'lib/hooks/useDidMount';
 import { useLocation } from 'react-router-dom';
 import { useNavbar } from 'lib/hooks/useNavbar';
+import { useZnsSdk } from 'lib/hooks/sdk';
 
 type ZNSProps = {
 	version?: number;
